create forecase_facade and test methods
