+++ app/controllers/api/v1/forecast_controller.rb
@@ -1,10 +1,9 @@
 class Api::V1::ForecastController < ApplicationController
 
   def show
-    @forecast = Forecast.new(params[:location])
-    @forecast.location
-    @forecast.get_forecast
-    render json: ForecastSerializer.new(@forecast)
+    @facade = ForecastFacade.new(params[:location])
+    @facade.get_forecast
+    render json: ForecastSerializer.new(@facade)
   end
 
 end
